% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{load_numbers}
\alias{load_numbers}
\title{Load Numbers}
\usage{
load_numbers(n_digit)
}
\arguments{
\item{n_digit}{\code{[numeric]} A number specifying type of table to be imported}
}
\value{
A data frame object
}
\description{
Import pre-computed data table
}
