% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{lucky_number_generator}
\alias{lucky_number_generator}
\title{Lucky Number Generator}
\usage{
lucky_number_generator(n_digit, ...)
}
\arguments{
\item{n_digit}{\code{[numeric]} A number specifying type of table to be imported}

\item{...}{\code{[expression]} Expression specifying condition to filter numbers}
}
\value{
A data table object
}
\description{
Generate numbers based upon pre-defined patterns
}
